requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.1
Adafruit-PureIO==0.2.1
Adafruit-SSD1306==1.6.1
Django==1.10.5
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
Werkzeug==0.9.6
argparse==1.2.1
backports.ssl-match-hostname==3.5.0.1
blinker==1.3
chardet==2.3.0
colorama==0.3.2
configparser==3.5.0
decorator==3.4.0
django-bootstrap3==8.2.1
django-extensions==1.7.6
django-tables2==1.4.2
funcsigs==1.0.2
gpiozero==1.3.1
hiredis==0.2.0
html5lib==0.999
ipython==2.3.0
itsdangerous==0.24
lxkeymap==0.1
mcpi==0.1.1
meld==3.12.1
meld3==1.0.0
mock==2.0.0
ndg-httpsclient==0.3.2
numpy==1.8.2
paho-mqtt==1.2
pbr==1.10.0
pexpect==3.2
picamera==1.12
picraft==1.0
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.30
pushbullet==0.1
py==1.4.32
pyOpenSSL==0.13.1
pyasn1==0.1.7
pygame==1.9.2a0
pygobject==3.14.0
pyinotify==0.9.4
pyserial==2.6
pytest==3.0.6
python-apt==0.9.3.12
python-magic==0.4.13
redis==2.10.5
requests==2.4.3
sense-emu==1.0
sense-hat==2.2.0
simplegeneric==0.8.1
six==1.8.0
smbus==1.1
spidev==3.0
stevedore==1.20.0
supervisor==3.0
twython==3.1.2
urllib3==1.9.1
urwid==1.2.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
websocket-client==0.40.0
wheel==0.24.0
wicd==1.7.2.4
wsgiref==0.1.2
